Added handling of empty collections
Moved collector options directly in the module.
Added a nodata call in the rating processor which is called every
time the collection process retrieves no data from the backend.
It can be used to process some billing calculations even without
data.
Added a nodata call in the storage backend to handle the empty
collections more efficiently.

Change-Id: I37430823d2a5b416c9b171dfce75e136b9fa40df