Renaming billing to rating
As CloudKitty is a rating and not a billing component, we needed to
rename all components and change paths on the API.
We've created a compatibility layer to help people with the migration.
The old API objects are now deprecated and shouldn't be used, they will
be removed in a near future.

Change-Id: I9e264f4ed01f4c94366eb51f612d634312b6a2f7