Added support for an hybrid gnocchi storage
This storage removes the need to duplicate the data present in gnocchi.
Every dataframe is associated to the metric_id that generated the
calculations.
The downside is that you can only use gnocchi with this storage driver.
Moved around some utility function to remove the need to duplicate code
and have a cleaner boundary between the different parts of the code.

Change-Id: Iabfdcd4c15c906ed145ce383b65d1538f72671aa