The result of tenant list may be unpredictable
A problem like bug 1512560 existed in orchestrator when determinate
which tenants will be rated in the next period:
The process method in orchestrator include the following processing
logic: modify (remove some items) the sequence (self._tenants) being
iterated over in the loop, which may cause unpredictable result.
This patch fixed this problem as the same way as solving bug 1512560.

Change-Id: Id148c5a0a1ab0290cfff02d9121badcb7f4e6ac7
Closes-Bug: #1520068