remove rating no match case
The no match case currently we write actually will never be invoked,
because there is no such mapping or threshold level named '_DEFAULT_',
it is actually a default group name.

And there is no need to worry about the no match case, if there is
no field or threshold matches, just ignore it and the rating will
be 0, because they will be ignored when we add rating info for collected
data.

Change-Id: I30e32d0a7d945ba1a5d5753fdb94b37b019d8e95