remove setting for option verbose
ck-api log complains:

Option "verbose" from group "DEFAULT" is deprecated for removal.
Its value may be silently ignored in the future.

If this option is not set explicitly, there is no such warning.
Furthermore, the default value of verbose is true, so there is
no need to set this value in devstack plugin.

Change-Id: I0dbf2ec729482883c87a89f2a13a406ff4998754