Fix issues with alembic constraint naming
Most constraint where unnamed in CloudKitty, mainly because the default
naming convention doesn't do so. Since most primary key and unique are
automatically defined by attributes on a column you can't directly
specify a name.
We've defined a naming convention in a base model to ensure new tables
are created with constraint having sensible names.
This still won't fix Alembic SQLite migrations that can't reproduce the
ON DELETE attribute of the ForeignKey. Creating all sorts of bugs with
delete recursivity.

Change-Id: I5f0cc1771082e4c53ad678d94fdf68af1323050d