Ensure module list is up to date in API tests
Sometimes rating modules are not loaded in memory. This seems to come
from the fact that a same thread is used for gabbi tests and only one
migration is applied for each test and the module list is created on
first load. We ensure that we reload module list between gabbi tests.

Change-Id: I767fd2aa1bd38f9af0c19123cc9738c9f4e57c40
(cherry picked from commit 532dd50333e6f339e8e0488ac9c5fe661b6718d5)