Fix incorrect rating for network.floating
The volume of floating.ip pushed by ceilometer (whatever backend
used) is the numerical status of the floating, leading to an
incorrect rating by CK.

With ceilometer collector, the value is already forced to 1 (as
for compute). This patch fixes the bug for the gnocchi collector.

An upper constraint has been set on oslo.config to match the ocata release.
(Some function prototypes have been modified,
see https://review.openstack.org/#/c/503181/)

Change-Id: I974db35a87c3a791f25e19dbbeac77a4fae0a25f
Closes-Bug: #1659491
Closes-Bug: #1722847