Ensure compatibility with all versions of gnocchiclient.
This makes sure that the gnocchi storage backend is compatible with all
versions of gnocchiclient. Starting with version 5.0.0 (which is not stable
yet), it returns python datetime objects instead of timestamps
(see https://github.com/gnocchixyz/python-gnocchiclient/commit/c14f5a484dd00c14854f56150fca5eee5c42a8b7).

As soon as a gnocchiclient version > to 4.0 is considered stable, the try/except
block can be removed and a constraint python-gnocchiclient>=5.0 may be set in
requirements.txt.

Change-Id: If1816b6926e8048ff05908a15fc0870050a5718a