Fix UsageEnd in CSV reports
This fixes the value of UsageEnd in CSV reports. Currently, it is always
equal to the usage_start of the first processed dataframe + one collect period.

It was an issue with the first condition of _get_usage_end(): given that
_get_usage_start() was always called before _get_usage_end(), the condition
was always true and cached_end was never updated.

Change-Id: Ice8f3b291a2b1327b9b95e1974de1431a93fd29e
Story: 2001414
Task: 6112