Update fields of CSV reports.
Gnocchi being the default collector, the metadata field to check for ItemName
has been renamed from 'flavor' to 'flavor_name' ('flavor_name' is a required
attribute for 'instance' resources in gnocchi).

The ResourceId field is now available for all services, through the
'resource_id' attribute.

An 'ItemFlavorId' field has been added.

The RateId and ItemDescription fields have been removed, because the only key of
the 'rating' dict is 'price'.

Change-Id: I5aed4b813940effa5e0b14ba6a4eca0d16ac0c40