Deprecate the ceilometer collector
The Ceilometer API has been deprecated for a while now. Some of the meter
types queried by the collector (instance, image) have been removed from the
Telemetry project in the Ocata release.

In order to make sure that the warning will be seen, it has been put in the
constructor of the Worker class.

We should completely remove the collector during the next release cycle.

Change-Id: Ieaa2bc585d0cbcf7b035da627e189c920c9a8e08