Make build reproducible
Currently in master, cloudkitty is leaking the build hostname in the
cloudkitty.conf file when generating it. This makes the entire package
no reproducible.

This patch adds a sample_default= argument when defining host=, so that
the default value when generating the config file is *not* the hostname
of the build host. This doesn't change at all the behavior, which will
use the hostname FQDN if the value isn't set by the user.

Change-Id: I59eb50018dd895f40ff949123c4ab78360c74850