Handle the scope id as a regular groupby attribute in storage
Until now, the scope_id was handled separately in the storage backend.
Given that the state is not handled by the storage backend anymore, the
scope_id should not be treated differently by the storage backend. This aims
at simplifying cloudkitty's internal architecture.

In order to make the scope available to the whole orchestrator/API context,
a "scope_key" option has been added to cloudkitty.conf, and it has been
removed from the extra_args from the Gnocchi and Monasca collectors.

Change-Id: Idaeb783cf6fb566d795b1a502f45d9260544ce02
Story: 2001372
Task: 26765