Don't update the state of a scope when no data was collected
This makes the worker return instead of raising NoDataCollected
when an exception occurs during collection.

Before this patch, a NoDataCollected exception was raised when a collect
exception occured. This caused the worker to update the collect state and
to process the next period, leading to missing data in the storage backend.
With this patch, the collect state is not updated anymore.

Change-Id: I0f7c6d6e4cdc28ad6245b863d396c4f9403de380
Story: 2003828
Task: 26761
Task: 26763