Don't heartbeat manually in the cloudkitty orchestrator
This fixes cloudkitty-processor randomly crashing when releasing locks.

Until now, heartbeating was done manually after each collect period,
once all tenants have been processed. However, the first collect period
is generally much longer than the following ones, given that the tenant workers
need to catch up with the current date.
Because of that, too much times passes between the moment where the coordinator
is created and the first heartbeat. This causes locks to be automatically
released, which leads to an error when we try to release them manually.

This issue is fixed by letting the tooz coordinator taking care of
heartbeating.

Change-Id: If82f5981645032dcc649ea6e29cfdf8b1054b01e
Story: 2004270
Task: 27817