Make devstack-tempest job voting
Since the zuul jobs have been converted to the new format (see commit de4a38e),
the devstack-tempest job is stable. In consequence, this makes this job voting.
Given that the "cloudkitty-devstack-install" job was the same as the
devstack-tempest job without the tempest run, it has been removed in order to
put less load on the gate.

Change-Id: I942f9069eedce033b0b087992f8b84c76149ff1a