Add DataPoint/DataFrame objects
This introduce the DataPoint and DataFrame objects. These are replacing the
dicts that were used until now:

A DataPoint is an immutable object representing one measure for one metric,
with associated price, qty, groupby and metadata.

A DataFrame is a collection of DataPoints, with helper functions for
point insertion and JSON formatting/loading.

Story: 2005890
Task: 35658
Change-Id: I71e95bba0a0cdd049b0fba3e79c7675b6365c37f