Fix StateManager.set_state() logic
This fixes the logic of the StateManager's set_state() method: if, by
accident, the state of an identifier is set to the same timestamp as it
currently is at, we won't attempt to create another object.

The 'uq_cloudkitty_storage_states_identifier' constraint was preventing a bug
from happening, but this fixes the code logic.

Change-Id: I530e63dbd67c1357d4aac964fffcdb45ce495768