Fix GET /v1/dataframes endpoint
When switching to DataFrames objects, a mistake has been made in the
DataFrameController object: the "resources" list was not emptied once a
dataframe was created, leading to a datapoint duplication in the command
output.

The bug hasn't been detected at that time, because each usage type in a
dataframe only contained a single data point. Gabbi fixtures have been
updated to include several points per dataframe and avoid a future
regression.

Change-Id: Ib41f0c91232ed06b82b6bb9f60ab316fc76e5209