Raise an exception in case of an invalid configuration file
Until now, the "load_conf" function did return an empty dict in case of
an invalid config. This was done because this function was called on
module import, leading to crashes during tox jobs.

However, catching the exception could lead to hard to debug issues.

Since the function isn't called on module import anymore, it does now
raise an exception in case the metrics.yml config file can't be read or has an
invalid format.

Story: 2006941
Task: 37631

Change-Id: Ia09ea8cacb7a82a4baf92f310f2428e2ec5973b7