Standardize aggregation methods and granularities for Gnocchi collector
This PR proposes the standardization of aggregation methods allowed by
CloudKitty in Gnocchi collector as the aggregation methods available in
Gnocchi API for both archive policies and aggregation API. The rationale is
that we need to support the same aggregation methods available there.
Otherwise, users can try to use something available in the Gnocchi API, and it
does not work due to a limitation in CloudKitty.

We also propose to set the default granularity as 3600. CloudKitty is already
using one hour time frames by default. There is no sense in not setting the
granularity as 3600 as well. If we do not define the granularity 3600 by
default, CloudKitty will always retrieve (in the aggregation API) measurements
for all available granularities, which is an overhead. Besides being an
overhead, it can cause huge inconsistencies, if for some reason (that we
painfully discovered in production), the granularity of 3600 is not returned.
The method "_format_data"  does not check if the data being
"formated"/obtained from the measurements object represents the 3600 (1h) time
frame, it just retrieves the first element, and moves on; thus, leading to
wrong billing information.

And last, but not least, we propose a debug log to show in the log files, when
debug log level is enabled, the response received in the aggregation API. This
can enable debug for operation people. Otherwise, we need to change code to be
able to troubleshoot problems with CloudKitty processing when using Gnocchi
collector.

Change-Id: I8a1a75a74e04cb9babdf09f115620b152861e218