Fix processing of "rate:xxx" re-aggregation configuration
The aggregation API in Gnocchi does not return anything for Granularity
3600, when the time-range is just one hour for "rate:xxx"
re-aggregation methods. That behavior on Gnocchi side is correct. The
"rate:xxx" aggregation method needs at least two data points (we are
taking the diff, right?). Therefore, when configured to work with
"rate:xxx" CloudKitty needs to call the aggregation API with a two
hours interval (for granularity 3600), and not just one.

With this patch, when CloudKitty detects a "rate:xxx" configuration, it
will use the range as two times (*) the granularity we are using to
process data. Therefore, if the granularity is one hour, it uses a
two-hour time frame; on the other hand, if the granularity is
configured to 10 minutes, it would consider a 20 minutes timeframe. I
am doing that to retrieve two data points at a time in the aggregation
API; thus, we can compute the diffs between the data points.

Depends-On: https://review.opendev.org/#/c/704787/
Change-Id: I570db3db4c969042038a286829c8df2ee06fb5d7