Make Gnocchi connection pool configurable
When CloudKitty is collecting data from Gnocchi, it opens many
connections to Gnocchi to parallelize the requests; as CloudKitty starts
to fetch the responses from Gnocchi, it releases the connections to the
pool until it starts to overflow (reaching the default limit of the
Requests connection pool). This behavior causes no problems, but
prevents the reuse of connections.

To mitigate this problem and optimize the use of connections, we propose
to add a new configuration in Gnocchi collector and fetcher to set the
connection pool maximum size when connecting to Gnocchi. The new
configuration is `http_pool_maxsize` and you can configure it in the
`cloudkitty.conf` file like:

```
[fetcher_gnocchi]
http_pool_maxsize = 50

[collector_gnocchi]
http_pool_maxsize = 50
```

The default value is defined by the `requests` library which is 10.

Change-Id: I3ee22984dc87b505924e1711bc723bc1c5f160a5
Story: 2008125
Task: 40847