Adjust hacking tests to fix py38 support
Starting in py38, some of the column calculations have changed. We had
some hacking unit tests that would assert on tuples containing those
column numbers, so things would pass fine when running on py36 or py37,
but would fail when run on py38.

Since we don't really care about the column number, only whether the
hacking check fails or doesn't, we can just match on mock.ANY for that
value.

Change-Id: I117177a43293efb389dd8c0cd5a5c0ee16a7eca9
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>