Add support for multiple value filters
The goal of this patch is to introduce support for multi-valued
parameters. For instance, for the `type` parameter, even though the code
was treating it as a possible list of types, the API would not allow a
user to send multiple types.

This patch enables users to send filters with multiple values, which can
be useful for filtering by project_ids for instance, or different types
(metric types).

Change-Id: I59397b33d014709eb976c78d517f009b8a2be4cf