SQLalchemy not creating constraint for Enum on version 1.4.0+
SQLachemy was upgraded via the `requirements` projects commit
(https://github.com/openstack/requirements/commit/dc86260b283dedc3076d7873f5f031f45e3e3671).

That upgrade broke some code. Therefore, this patch intends to fix
the code that was broken due to a method removal and also due to a
change in `create_constraint` for Enum types, which was defaulted
to `False` in version 1.4.0

Change-Id: Ib36a7275ce8ff30aa19cbde521fd37f60bd00ff8