Validates the period compatibility of reprocessing dates
Problem description
===================

The reprocess API is accepting time windows that are
not compatible with the configured collection period
which causes some reprocessings to be endless or
generating different values based on the time window
the user inputs.

Proposal
========

We propose to add a validation in the reprocess API
to deny users to schedule a reprocess using a not
compatible time window and suggest the nearest valid
time window that the user can schedule a reprocess.

Change-Id: I24745a612bbd4714a7793df1deced671c1d1c26a