Add missing api-ref envlist to tox
This is a follow-up to:
https://review.opendev.org/c/openstack/cloudkitty/+/867122

We need to add the `[testenv:api-ref]` section to `tox.ini` to be able
to run `tox -e api-ref`, which is what Zuul uses to build the API refs.

Sphinx had to be capped to the latest version 4 or otherwise `tox` fails
with the "add_content() takes 2 positional arguments but 3 were given"
error message.

Change-Id: I65b008152e2bc64f29229996c87ad4587fe85043