Introduce start and end dates on rating rules
This change affects both hashmap and pyscripts. This feature adds a time
to live for each rating rule. The change is also replacing the deletion
process from removing the DB rows to marking them as deleted, allowing
users to know when and who deleted a specific rule. Name and description
metadata were added to enrich the rules with more context of their
creation.

Change-Id: Icac45c8f3ac8b5d86a134b311de9a1a77932b003
Depends-On: https://review.opendev.org/c/openstack/cloudkitty-tempest-plugin/+/892382
Signed-off-by: Pedro Henrique <phpm13@gmail.com>