Fix retrieval of reprocessing tasks
Calling GET /v2/task/reprocesses with python-cloudkittyclient was
returning Internal Server Error, with the following API trace:

      File "/var/lib/kolla/venv/lib/python3.6/site-packages/cloudkitty/api/v2/task/reprocess.py", line 259, in get
        order, ACCEPTED_GET_REPROCESSING_REQUEST_ORDERS)
    TypeError: __init__() takes from 1 to 3 positional arguments but 4 were given

This was because http_exceptions.BadRequest was given multiple arguments
(similar to LOG.* methods) instead of a single string.

Another issue is that python-cloudkittyclient sends the "DESC" order
while the API only supports "desc" and "asc". Convert to lower case for
compatibility.

Change-Id: Id1145adff82bc9a01e4eb0f306f0bfa535142459