Allow Multiple Ratings for the same Metric on Prometheus
This commit moves a part of the gnocchi collector to the base collector,
so that all collectors can benefit from multirating per metric.

It solves the issue with Prometheus having one "*_info" metric containing
the scope and multiple other metric containing the values we want to rate.

Also splitting prometheus' fetch_all, to gain in readability and allow
us to render the promql query from tests.

Change-Id: I4f0892a80f37ce052f573e7622534aa728e7cff1