Add handling to metric collect error
As we were simply exiting on metric collect for a scope
when any unhandled error happens, other scopes that
should be processed after that problematic scope will
not be processed, and if the problem persist with this
specific scope, it will make CloudKitty to never move
on to the next scopes as the processor will stop
processing.

This proposal is to raise an exception instead of killing
the processor when an error happens in the metric collect,
so the problematic scope will be skipped till next
collection and the other scopes will be processed normally.

Story: 2011321
Task: 51519
Change-Id: I460877e818166230014c1d1ae7d9588162554264