Deploy healthcheck middleware as app
Using the healthcheck middleware as a filter is deprecated and
the middleware should be used as an application[1].
 [1] 6feaa13610c450c8486f969703768db5319b4846

This change updates definition and usage of the healthcheck middleware
accordingly to avoid the following deprecation warning.

DeprecationWarning: Using function/method 'Healthcheck.factory()' is
deprecated: The healthcheck middleware must now be configured as
an application, not as a filter.

Change-Id: Idf31c0ed5f433e42d26560d5d0022e53faeb6450
Signed-off-by: Takashi Kajinami <kajinamit@oss.nttdata.com>