Shard the requests to Loki
Loki default options are very conservative about resourcing and
access so the queries are manageable for the user. This patch
introduced sharding of queries so the user can decide to do more
queries to Loki (with the corresponding time overhead) in exchange
for being able to calculate and retrieve more granular data.

The user can also decide to store less granular data into Loki and
do less query sharding, improving the response time.

Change-Id: I309025e2e4c914aeb593cff7d88469a351edf2fd
Signed-off-by: Juan Larriba <jlarriba@redhat.com>