Add reject_old_samples_max_age to loki config
By default Loki allows to push samples at most 1 week old. When
cloudkitty processor is started for the first time, it'll try
to process data from the 1st day of the current month and store
them into the configured storage backend. If it's currently after
the 7th day of the month and the collector has data older than a
week, the processor will try to push data to Loki older than
a week, which Loki will reject by default.

This change configures Loki deployed in devstack to accept data
up to 31 days old.

Change-Id: Ieb43692f9a4f84b1b298fbaa4fdb736544f85b64
Signed-off-by: Jaromir Wysoglad <jwysogla@redhat.com>