Fix Loki storage delete with None end parameter
The reset_state method in the orchestrator calls storage.delete() with
end=None to delete data from a given timestamp to the present. The Loki
client's delete_by_query method was not handling None for the end
parameter, causing an AttributeError when trying to call timestamp() on
None.

This change adds a check to default end to the current UTC time when
it is None, consistent with how other storage backends handle this case.

Change-Id: Ie3b6be09d0d70a79d35e16c23faf091f4dea9f6d
Signed-off-by: Juan Larriba <jlarriba@redhat.com>